requirements.txt
python-fasthtml
SQLAlchemy
uvicorn
openai
jinja2
llm
